Removed redundant code, moved some file functions to the files library
