+++ pkg/shared/utils.go
@@ -14,15 +14,6 @@ import (
 	"strings"
 	"text/template"
 
-	crssh "golang.org/x/crypto/ssh"
-
-	"github.com/gitsight/go-vcsurl"
-	"github.com/go-git/go-git/v5"
-	"github.com/go-git/go-git/v5/config"
-	"github.com/go-git/go-git/v5/plumbing"
-	"github.com/go-git/go-git/v5/plumbing/transport"
-	"github.com/go-git/go-git/v5/plumbing/transport/http"
-	"github.com/go-git/go-git/v5/plumbing/transport/ssh"
 	"github.com/hashicorp/go-hclog"
 )
 
@@ -64,156 +55,6 @@ func ExpandPath(path string) (string, error) {
 		return filepath.Join(homeDir, path[2:]), nil
 	}
 	return path, nil
-}
-
-func getGitAuth(args *VCSFetchRequest, variables *EvnVariables, logger hclog.Logger) (transport.AuthMethod, error) {
-	var auth transport.AuthMethod
-	var err error
-
-	switch args.AuthType {
-	case "ssh-key":
-		logger.Debug("Setting up SSH key authentication")
-
-		// Handle paths starting with tilde, like ~/.ssh/id_rsa
-		// https://gist.github.com/miguelmota/9ab72c5e342f833123c0b5cfd5aca468?permalink_comment_id=3953465#gistcomment-3953465
-		if strings.HasPrefix(args.SSHKey, "~/") {
-			dirname, _ := os.UserHomeDir()
-			args.SSHKey = filepath.Join(dirname, args.SSHKey[2:])
-		}
-
-		if _, err := os.Stat(args.SSHKey); err != nil {
-			logger.Error("Reading file with a key is failed ", "path", args.SSHKey, "error", err.Error())
-			return nil, err
-		}
-
-		auth, err = ssh.NewPublicKeysFromFile("git", args.SSHKey, variables.SshKeyPassword)
-		if err != nil {
-			logger.Error("An extraction publickeys process is failed", "error", err.Error())
-			return nil, err
-		}
-
-		auth.(*ssh.PublicKeys).HostKeyCallbackHelper = ssh.HostKeyCallbackHelper{
-			HostKeyCallback: crssh.InsecureIgnoreHostKey(),
-		}
-
-	case "ssh-agent":
-		logger.Debug("Setting up SSH agent authentication")
-
-		auth, err = ssh.NewSSHAgentAuth("git")
-		if err != nil {
-			logger.Error("Failed to create SSH agent auth", "err", err)
-			return nil, err
-		}
-
-		auth.(*ssh.PublicKeysCallback).HostKeyCallbackHelper = ssh.HostKeyCallbackHelper{
-			HostKeyCallback: crssh.InsecureIgnoreHostKey(),
-		}
-
-	case "http":
-		logger.Debug("Setting up HTTP authentication")
-		auth = &http.BasicAuth{
-			Username: variables.Username,
-			Password: variables.Token,
-		}
-
-	default:
-		err := fmt.Errorf("Unknown auth type: %s", args.AuthType)
-		logger.Error("Problems with a git fetching process", "error", err)
-	}
-	return auth, err
-}
-
-func GitClone(args VCSFetchRequest, variables EvnVariables, logger hclog.Logger) (string, error) {
-	info, err := vcsurl.Parse(args.CloneURL)
-	if err != nil {
-		logger.Error("Unable to parse VCS url", "VCSURL", args.CloneURL, "err", err)
-		return "", err
-	}
-
-	branch := plumbing.ReferenceName(args.Branch)
-	auth, err := getGitAuth(&args, &variables, logger)
-	if err != nil {
-		logger.Error("Unable to set up auth", "err", err)
-		return "", err
-	}
-
-	// debug output from git cli
-	output := logger.StandardWriter(&hclog.StandardLoggerOptions{
-		InferLevels: true,
-		ForceLevel:  hclog.Debug,
-	})
-
-	gitCloneOptions := &git.CloneOptions{
-		Auth:          auth,
-		URL:           args.CloneURL,
-		ReferenceName: branch,
-		Progress:      output,
-		Depth:         1,
-	}
-
-	gitPullOptions := &git.PullOptions{
-		Auth:          auth,
-		ReferenceName: branch,
-		Progress:      output,
-		Depth:         1,
-	}
-
-	gitCheckoutOptions := &git.CheckoutOptions{
-		Branch: branch,
-		Force:  true,
-	}
-
-	gitFetchOptions := &git.FetchOptions{
-		RemoteName: "origin",
-		Auth:       auth,
-		Progress:   output,
-		RefSpecs:   []config.RefSpec{config.RefSpec("+refs/*:refs/*")},
-		Depth:      1,
-	}
-
-	logger.Debug("Fetching repo", "repo", info.Name, "branch", args.Branch, "targetFolder", args.TargetFolder)
-	_, errClone := git.PlainClone(args.TargetFolder, false, gitCloneOptions)
-	if errClone != nil && errClone != git.ErrRepositoryAlreadyExists {
-		logger.Error("Error on Clone occured", "err", err, "targetFolder", args.TargetFolder, "remote", gitCloneOptions.URL)
-		return "", err
-	}
-
-	r, err := git.PlainOpen(args.TargetFolder)
-	if err != nil {
-		logger.Error("Can't open repository on a disk", "err", err, "targetFolder", args.TargetFolder)
-		return "", err
-	}
-
-	_ = r.Fetch(gitFetchOptions) // updating list of references
-
-	w, err := r.Worktree()
-	if err != nil {
-		logger.Error("Error on Worktree occured", "err", err, "targetFolder", args.TargetFolder)
-		return "", err
-	}
-
-	logger.Debug("Checkout a branch", "repo", info.Name, "targetFolder", args.TargetFolder, "branch", args.Branch)
-	if err = w.Checkout(gitCheckoutOptions); err != nil {
-		logger.Error("Error on Checkout occured", "err", err, "targetFolder", args.TargetFolder)
-		return "", err
-	}
-
-	logger.Debug("Reseting local repo", "repo", info.Name, "targetFolder", args.TargetFolder)
-	if err := w.Reset(&git.ResetOptions{Mode: git.HardReset}); err != nil {
-		fmt.Println("Error on Checkout occured", "err", err, "targetFolder", args.TargetFolder)
-		return "", err
-	}
-
-	if errClone != nil && errClone == git.ErrRepositoryAlreadyExists {
-		logger.Debug("Pulling repo", "repo", info.Name, "targetFolder", args.TargetFolder, "branch", args.Branch)
-		if err = w.Pull(gitPullOptions); err != nil {
-			logger.Error("Error on Pull occured", "err", err, "targetFolder", args.TargetFolder)
-			return "", err
-		}
-	}
-
-	logger.Info("A fetch function finished", "repo", info.Name, "branch", args.Branch, "targetFolder", args.TargetFolder)
-	return args.TargetFolder, nil
 }
 
 func ExtractRepositoryInfoFromURL(Url string, VCSPlugName string) (string, string, string, string, string, string, error) {
