+++ plugins/github/github.go
@@ -10,10 +10,12 @@ import (
 	"github.com/hashicorp/go-hclog"
 	"github.com/hashicorp/go-plugin"
 
+	"github.com/scan-io-git/scan-io/internal/git"
 	"github.com/scan-io-git/scan-io/pkg/shared"
 	"github.com/scan-io-git/scan-io/pkg/shared/config"
 )
 
+// VCSGitlab implements VCS operations for Gitlab.
 type VCSGithub struct {
 	logger       hclog.Logger
 	globalConfig *config.Config
@@ -71,16 +73,9 @@ func (g *VCSGithub) AddCommentToPR(args shared.VCSAddCommentToPRRequest) (bool, 
 func (g *VCSGithub) Fetch(args shared.VCSFetchRequest) (shared.VCSFetchResponse, error) {
 	var result shared.VCSFetchResponse
 
-	//variables, err := g.init("fetch")
-	variables := shared.EvnVariables{}
-	// if err != nil {
-	// 	g.logger.Error("Fetching is failed", "error", err)
-	// 	return err
-	// }
-
-	path, err := shared.GitClone(args, variables, g.logger)
+	path, err := git.CloneRepository(g.logger, g.globalConfig, &args)
 	if err != nil {
-		g.logger.Error("A fetching function is failed", "error", err)
+		g.logger.Error("failed to clone repository", "error", err)
 		return result, err
 	}
 	result.Path = path
